No need to aggregate coverage for local experiments

+++ run_all_experiments.py
@@ -547,9 +547,12 @@ def main():
   # Set global variables that are updated throughout experiment runs.
   WORK_DIR = args.work_dir
 
-  coverage_gains_process = Process(
-      target=extend_report_with_coverage_gains_process)
-  coverage_gains_process.start()
+  if args.cloud_experiment_name:
+    coverage_gains_process = Process(
+        target=extend_report_with_coverage_gains_process)
+    coverage_gains_process.start()
+  else:
+    coverage_gains_process = None
 
   experiment_results = []
   if NUM_EXP == 1:
@@ -575,9 +578,10 @@ def main():
       # Wait for all workers to complete.
       p.join()
 
-  # Do a final coverage aggregation.
-  coverage_gains_process.kill()
-  extend_report_with_coverage_gains()
+  if coverage_gains_process:
+    # Do a final coverage aggregation.
+    coverage_gains_process.kill()
+    extend_report_with_coverage_gains()
 
   # Capture time at end
   end = time.time()
